+++ backend/cmd/meguri-cli/main.go
@@ -10,6 +10,7 @@ import (
 	// プラグインのinit処理の実行
 	_ "meguri/plugins/fetcher-chromium"
 	_ "meguri/plugins/fetcher-http"
+	_ "meguri/plugins/filter-exclude-selectors"
 	_ "meguri/plugins/filter-maincontent"
 	_ "meguri/plugins/filter-selector"
 	_ "meguri/plugins/linkextractor-default"