+++ backend/internal/usecase/scrape_cache_test.go
@@ -12,6 +12,7 @@ import (
 
 	_ "meguri/plugins/fetcher-chromium"
 	_ "meguri/plugins/fetcher-http"
+	_ "meguri/plugins/filter-exclude-selectors"
 	_ "meguri/plugins/filter-maincontent"
 	_ "meguri/plugins/filter-selector"
 	_ "meguri/plugins/linkextractor-default"